558: Add function MapDocument::createLinkedGroups
This function creates a new member of the link set of the currently
selected group and adds it to the document.

When a group is added to the document, it must be linked to its link
set, and it must be unlined when it is removed from the document. We add
the necessary code and some tests.